--- services/auth-service/src/main/java/com/ifortex/internship/auth_service/dto/TokenRefreshResponse.java
@@ -1,14 +1,0 @@
-package com.ifortex.internship.auth_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenRefreshResponse {
-  private String accessToken;
-}
