+++ services/auth-service/src/main/java/com/ifortex/internship/auth_service/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.ifortex.internship.auth_service.dto;
+package com.ifortex.internship.auth_service.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
