+++ services/auth-service/src/main/java/com/ifortex/internship/auth_service/dto/response/AuthResponse.java
@@ -1,0 +1,17 @@
+package com.ifortex.internship.auth_service.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class AuthResponse {
+  private CookieTokensResponse cookieTokensResponse;
+  private Long userId;
+}
