+++ services/auth-service/src/main/java/com/ifortex/internship/auth_service/dto/response/CookieTokensResponse.java
@@ -1,0 +1,16 @@
+package com.ifortex.internship.auth_service.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.http.ResponseCookie;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CookieTokensResponse {
+  private ResponseCookie accessCookie;
+  private ResponseCookie refreshCookie;
+}
